:art: restructure geometry module
