+++ mindee/geometry/error.py
@@ -1,0 +1,2 @@
+class GeometryError(RuntimeError):
+    """An error related to geometry operations."""
