--- mindee/http/error.py
@@ -1,1 +1,0 @@
-# TODO
