--- mindee/parsing/common/mvision_v1.py
@@ -1,1 +1,0 @@
-# TODO
