--- mindee/parsing/common/ocr_page.py
@@ -1,1 +1,0 @@
-# TODO
