--- mindee/parsing/common/page.py
@@ -1,1 +1,0 @@
-# TODO
