--- mindee/parsing/common/summary_helper.py
@@ -1,1 +1,0 @@
-# TODO
