+++ mindee/product/custom/__init__.py
@@ -1,1 +1,1 @@
-from .custom_v1 import CustomV1, TypeCustomV1
+from mindee.product.custom.custom_v1 import CustomV1, TypeCustomV1
