+++ mindee/product/invoice_splitter/__init__.py
@@ -1,1 +1,4 @@
-from .invoice_splitter_v1 import InvoiceSplitterV1, TypeInvoiceSplitterV1
+from mindee.product.invoice_splitter.invoice_splitter_v1 import (
+    InvoiceSplitterV1,
+    TypeInvoiceSplitterV1,
+)
