+++ mindee/product/receipt/__init__.py
@@ -1,4 +1,4 @@
-from .receipt_v4 import ReceiptV4, TypeReceiptV4
+from mindee.product.receipt.receipt_v4 import ReceiptV4, TypeReceiptV4
 
-# from .receipt_v5 import ReceiptV5, TypeReceiptV5
-# from .receipt_v5_line_item import ReceiptV5LineItem
+# from mindee.product.receipt.receipt_v5 import ReceiptV5, TypeReceiptV5
+# from mindee.product.receipt.receipt_v5_line_item import ReceiptV5LineItem
